requirements.txt
Django==6.0
django-environ==0.12.0
pillow==12.0.0
